+++ lib/libsesame3bt-core/os3.cpp
@@ -147,7 +147,7 @@ OS3Handler::handle_history(const std::byte* in, size_t in_len) {
 	}
 	history.result = static_cast<Sesame::result_code_t>(in[0]);
 	if (history.result != Sesame::result_code_t::success || in_len < sizeof(Sesame::response_history_5_t)) {
-		DEBUG_PRINTLN("%u: Empty history", history.result);
+		DEBUG_PRINTLN("%u: Empty history", static_cast<uint8_t>(history.result));
 		client->fire_history_callback(history);
 		return;
 	}
